requirements.txt
requests==2.32.3
pandas==2.2.2
gspread==6.1.4
google-auth==2.34.0
